+++ AngularPrototype/angular/moduleBase/projects/jngcontrols-bootstrap3/src/controls/grid/gridbutton.ts
@@ -1,23 +1,48 @@
 import { Component, Input, Output, EventEmitter, ElementRef } from '@angular/core';
 import { NgGridCommon, NgGridButtonCommon } from '@jnetwork/jngcontrols-common';
 
-
+/**
+ * Grid Action Button
+ *
+ * @example Standard Action
+ *
+ * <ngGridButton icon="edit" (onclick)="action('edit')" [isdisabled]="true"></ngGridButton>
+ *
+ * @example Custom Action
+ *
+ * <ngGridButton iconstyle="sprite" icon="icon-sprite-base-main_info" (onclick)="action('info')"></ngGridButton>
+ *
+ */
 @Component({
   selector: 'ngGridButton',
   templateUrl: './gridbutton.html'
 })
 export class NgGridButton extends NgGridButtonCommon {
 
-
+  /**
+   * Gibt das Icon für den Button zurück
+   */
   public getIconClass(): string {
-    switch (this.iconstyle) {
-      case "edit":
-        return "sprite icon-sprite-base-main_edit";
-      case "delete":
-        return "sprite icon-sprite-base-main_delete";
-      default:
-        return this.iconstyle;
+    let iconset: string = this.iconstyle;
+    let iconcss: string = this.icon;
+
+    if (iconset === '') {
+      switch (this.icon) {
+        case "edit":
+          iconset = "sprite";
+          iconcss = "icon-sprite-base-main_edit";
+        case "delete":
+          iconset = "sprite";
+          iconcss = "icon-sprite-base-main_delete";
+      }
     }
+
+    if (this._isdisabledvalue)
+      iconcss += "_disabled";
+
+    let result = iconset + " " + iconcss;
+
+    return result.trim();
   }