+++ AngularPrototype/angular/moduleBase/projects/jngcontrols-common/src/controls/input/inputsearch.ts
@@ -5,12 +5,14 @@ import { NgInputCommon } from "./input";
 
 export class NgInputSearchCommon extends NgInputCommon {
 
-  @Input("iconname") _searchIconName: string = ""
+  @Input("iconname") _searchIconName: string = "";
+
+  @Input("buttontext") _buttontext: string = "";
 
   @Output("onclick")
   clickaction: EventEmitter<any> = new EventEmitter<any>();
 
-  searchClick(){
+  searchClick() {
     this.clickaction.emit(this.value)
   }
 }